requirements.txt
openai~=1.57.0

fastapi[standard]~=0.115.6
python-dotenv~=1.0.1
ruff
PyYAML~=6.0.2